server/requirements.txt
Django~=4.2.2
python-decouple==3.8
mysqlclient
djangorestframework~=3.14.0